requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==4.2.18
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==33.3.1
fonttools==4.55.3
kiwisolver==1.4.8
Markdown==3.7
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary>=2.9.9
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
validate-docbr==1.10.0
